Moved MapView to components folder
