+++ src/app/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Footer from './home/components/Footer';
 import Header from './home/components/Header';
-import MapView from './map/MapView';
+import MapView from './map/components/MapView';
 import global from './home/css/global.css'; // eslint-disable-line
 import Trivia from './trivia/components/Trivia';
 import BackToTop from './home/components/BackToTop';
